Pequeno ajuste na remoção em tab1, tab2 e tab3(Vigésimo segundo dia)

+++ app/src/app/pages/tab3/tab3.page.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Subscription } from 'rxjs';
+import { Subscription, Observable } from 'rxjs';
 import { Avaliacao } from 'src/app/interfaces/avaliacao';
 import { AvaliacaoService } from 'src/app/services/avaliacao.service';
+import { AlertController } from '@ionic/angular';
 
 @Component({
   selector: 'app-tab3',
@@ -15,6 +16,7 @@ export class Tab3Page implements OnInit {
   private avaliacaoSubscription: Subscription;
 
   constructor(
+    private alertController: AlertController,
     private avaliacaoService: AvaliacaoService,
     private router: Router
   ) {
@@ -30,9 +32,33 @@ export class Tab3Page implements OnInit {
   ngOnInit() {
   }
 
+  private presentAlert(): boolean | Promise<boolean> | Observable<boolean> {
+ 
+    return new Promise((resolve: any, reject: any) => {
+      this.alertController.create({
+        header: 'Tem certeza que deseja excluir essa avaliação?',
+        message: 'Clique ok para confirmar!',
+        buttons: [
+          {
+            text: 'notOk',
+            handler: _=> reject(false)
+        },
+        {
+            text: 'ok',
+            handler: _=> resolve(true)
+        }
+        ]
+      }).then(alert => alert.present());
+    });
 
-  removeAvaliacao(id: string){
-    return this.avaliacaoService.removeAvaliacao(id);
+  }
+
+
+  async removeAvaliacao(id: string){
+    let opc = await this.presentAlert();
+    if (opc){
+      return this.avaliacaoService.removeAvaliacao(id);
+    }   
   }
 
   pageAddAvaliacao() {
